+++ acp/acp-spring-boot-starter/src/main/java/pers/acp/springboot/core/conf/ControllerAspectConfiguration.java
@@ -2,6 +2,7 @@ package pers.acp.springboot.core.conf;
 
 import org.springframework.boot.context.properties.ConfigurationProperties;
 
+import java.util.ArrayList;
 import java.util.List;
 
 /**
@@ -11,6 +12,14 @@ import java.util.List;
 @ConfigurationProperties(prefix = "controller-aspect")
 public class ControllerAspectConfiguration {
 
+    public boolean isEnabled() {
+        return enabled;
+    }
+
+    public void setEnabled(boolean enabled) {
+        this.enabled = enabled;
+    }
+
     public List<String> getNoLogUriRegexes() {
         return noLogUriRegexes;
     }
@@ -19,6 +28,8 @@ public class ControllerAspectConfiguration {
         this.noLogUriRegexes = noLogUriRegexes;
     }
 
-    private List<String> noLogUriRegexes;
+    private boolean enabled = true;
+
+    private List<String> noLogUriRegexes = new ArrayList<>();
 
 }