+++ test/testkotlin/src/main/kotlin/com/test/kotlin/TestKotlinApplication.kt
@@ -1,0 +1,15 @@
+package com.test.kotlin
+
+import org.springframework.boot.autoconfigure.SpringBootApplication
+import org.springframework.boot.runApplication
+
+/**
+ * @author zhangbin by 28/04/2018 12:42
+ * @since JDK1.8
+ */
+@SpringBootApplication
+class TestKotlinApplication
+
+fun main(args: Array<String>) {
+    runApplication<TestKotlinApplication>(*args)
+}