+++ test/testspringboot/src/main/java/pers/acp/test/application/base/BaseRepository.java
@@ -2,6 +2,7 @@ package pers.acp.test.application.base;
 
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.data.jpa.repository.JpaSpecificationExecutor;
+import org.springframework.data.repository.NoRepositoryBean;
 
 import java.io.Serializable;
 
@@ -9,5 +10,6 @@ import java.io.Serializable;
  * @author zhangbin by 2018-1-15 16:59
  * @since JDK1.8
  */
+@NoRepositoryBean
 public interface BaseRepository<T, ID extends Serializable> extends JpaSpecificationExecutor<T>, JpaRepository<T, ID> {
 }