+++ doc/version_history.md
@@ -1,4 +1,18 @@
 ## 版本更新记录
+##### 4.0.0
+    1. 优化 gradle 脚本，spring cloud 版本号写入 dependencies.gradle；删除 cloud 模块下的 build.gradle
+    2. 升级 spring-cloud 版本为 Finchley.RC1
+    3. 优化 hystrix 断路监控
+    4. 优化 feign oauth 验证
+    5. 升级 spring-boot 版本为 2.0.1.RELEASE
+    6. 更新各依赖包版本
+    7. gateway 增加自定义断路器
+    8. 修改 Eureka Server 及 Client 配置，优化服务注册/发现感知度
+    9. 优化 Controller 切片日志，重命名切片类
+    10. kotlin 版本升级至 1.2.41，增加 kotlin demo，test:testkotlin
+    11. 集成 junit5 单元测试，并增加测试用例编写，test:testspringboot、test:testkotlin
+    12. 取消 specification-arg-resolver 集成
+    13. 增加 scala demo, test:testscala
 ##### 3.9.0
     1. 升级 spring-cloud 版本为 Finchley.M9
     2. spring cloud 网关服务由 zuul 更换为 spring-cloud-gateway